// backend/src/shared/dto/url.dto.ts
export interface UrlDto {
  id?: number;
  list?: string;
  tag?: string;
  group_transfer?: string;
  site_transfer?: string;
  search_site?: string;
  active?: boolean,
  publish?: boolean;
  tmp?: boolean;
  publishtime?: number;
  updatetime?: number;
  createtime?: number;
  user_id?: number;
}